refactor: :recycle: вынести project specs metadata
