+++ packages/core/metadata/importBoundaries.test.ts
@@ -32,8 +32,9 @@ const REGISTRATION_ENTRYPOINT_ALLOWLIST = new Set([
   "metadata/register.ts",
   "metadata/register.test.ts",
   "tests/setupTests.ts",
-  "metadata/validation/schemaRegistry.ts",
-  "metadata/validation/projectSpecs.ts",
+  "metadata/project/schemaRegistry.ts",
+  "metadata/project/specs.ts",
+  "metadata/project/specs.test.ts",
   "metadata/validation/validateForm.ts",
   "metadata/validation/dataPath/formTraversal.ts",
   "metadata/forms/clientApplicationForm/convertFromXML.ts",
