+++ packages/core/metadata/project/index.ts
@@ -1,0 +1,7 @@
+export {
+  configurationMetadataProjectSpec,
+  getMetadataProjectSpecByDir,
+  metadataProjectSpecByDir,
+  metadataProjectSpecs,
+  type MetadataProjectSpec,
+} from "./specs"
